+++ graphql/src/main/kotlin/com/sangsiklog/exception/HttpServiceException.kt
@@ -1,5 +1,5 @@
 package com.sangsiklog.exception
 
-class UnauthorizedException(
+class HttpServiceException(
     override val message: String
 ): RuntimeException(message)
