+++ graphql/src/main/kotlin/com/sangsiklog/service/user/UserService.kt
@@ -1,7 +1,7 @@
 package com.sangsiklog.service.user
 
 import com.sangsiklog.model.user.User
-import com.sangsiklog.service.HttpClient
+import com.sangsiklog.utils.http.HttpClient
 import com.sangsiklog.service.user.request.CreateUserRequest
 import com.sangsiklog.service.user.request.UpdateUserRequest
 import lombok.RequiredArgsConstructor
