+++ graphql/src/main/kotlin/com/sangsiklog/utils/http/ApiResponse.kt
@@ -1,7 +1,7 @@
-package com.sangsiklog.model
+package com.sangsiklog.utils.http
 
 data class ApiResponse<T> (
-    val code: Int,
+    val code: Int = 0,
     val message: String? = null,
     val result: T? = null
 )
